# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EduJobs Connect API",
    "description": "Comprehensive API for the EduJobs Connect platform - connecting Rwandan students with universities and job opportunities",
    "version": "1.0.0",
    "contact": {
      "name": "EduJobs Connect Support",
      "email": "support@edujobsconnect.rw",
      "url": "https://edujobsconnect.rw"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.edujobsconnect.rw/api",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Enter your token in the text input below."
      }
    },
    "schemas": {
...
      "University": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "objectId",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "example": "University of Rwanda"
          },
          "description": {
            "type": "string",
            "example": "Leading university in Rwanda offering quality education"
          },
          "type": {
            "type": "string",
            "enum": ["public", "private"],
            "example": "public"
          },
          "established": {
            "type": "integer",
            "example": 1963
          },
          "location": {
            "$ref": "#/components/schemas/UniversityLocation"
          },
          "contact": {
            "$ref": "#/components/schemas/UniversityContact"
          },
          "accreditation": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["HEC", "MINEDUC"]
          },
          "programs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Program"
            }
          },
          "rating": {
            "$ref": "#/components/schemas/UniversityRating"
          },
          "statistics": {
            "$ref": "#/components/schemas/UniversityStatistics"
          },
          "featured": {
            "type": "boolean",
            "default": false,
            "example": true
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive"],
            "default": "active",
            "example": "active"
          },
          "logo": {
            "type": "string",
            "format": "uri",
            "example": "/uploads/universities/ur_logo.png"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
      "UniversityLocation": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "KK 737 Street"
          },
          "city": {
            "type": "string",
            "example": "Kigali"
          },
          "province": {
            "type": "string",
            "example": "Kigali City"
          },
          "country": {
            "type": "string",
            "default": "Rwanda",
            "example": "Rwanda"
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "latitude": {
                "type": "number",
                "format": "float",
                "example": -1.9403
              },
              "longitude": {
                "type": "number",
                "format": "float",
                "example": 30.0594
              }
            }
          }
        }
      },
      "UniversityContact": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+250788123456"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "info@ur.ac.rw"
          },
          "website": {
            "type": "string",
            "format": "uri",
            "example": "https://ur.ac.rw"
          },
          "admissions": {
            "type": "string",
            "format": "email",
            "example": "admissions@ur.ac.rw"
          }
        }
      },
      "Program": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Computer Science"
          },
          "level": {
            "type": "string",
            "enum": ["bachelor", "master", "phd", "diploma", "certificate"],
            "example": "bachelor"
          },
          "duration": {
            "type": "integer",
            "description": "Duration in years",
            "example": 4
          },
          "description": {
            "type": "string",
            "example": "Comprehensive computer science program"
          },
          "requirements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["High school diploma", "Mathematics background"]
          },
          "tuition": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "number",
                "format": "float",
                "example": 1500000
              },
              "currency": {
                "type": "string",
                "default": "RWF",
                "example": "RWF"
              },
              "perYear": {
                "type": "boolean",
                "default": true,
                "example": true
              }
            }
          },
          "applicationDeadline": {
            "type": "string",
            "format": "date",
            "example": "2024-08-31"
          }
        }
      },
      "UniversityRating": {
        "type": "object",
        "properties": {
          "overall": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 5,
            "example": 4.2
          },
          "academic": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 5,
            "example": 4.5
          },
          "facilities": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 5,
            "example": 4.0
          },
          "campus": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 5,
            "example": 3.8
          },
          "totalReviews": {
            "type": "integer",
            "example": 1250
          }
        }
      },
      "UniversityStatistics": {
        "type": "object",
        "properties": {
          "totalStudents": {
            "type": "integer",
            "example": 25000
          },
          "internationalStudents": {
            "type": "integer",
            "example": 2000
          },
          "facultyMembers": {
            "type": "integer",
            "example": 1500
          },
          "studentFacultyRatio": {
            "type": "number",
            "format": "float",
            "example": 16.7
          },
          "graduationRate": {
            "type": "number",
            "format": "float",
            "example": 85.5
          }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "objectId",
            "example": "507f1f77bcf86cd799439011"
          },
          "title": {
            "type": "string",
            "example": "Software Developer"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          },
          "description": {
            "type": "string",
            "example": "We are looking for a skilled software developer..."
          },
          "requirements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Bachelor's degree in Computer Science", "3+ years experience", "JavaScript proficiency"]
          },
          "responsibilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Develop web applications", "Collaborate with team", "Code reviews"]
          },
          "employment": {
            "$ref": "#/components/schemas/Employment"
          },
          "location": {
            "$ref": "#/components/schemas/JobLocation"
          },
          "salary": {
            "$ref": "#/components/schemas/Salary"
          },
          "benefits": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Health insurance", "Remote work", "Professional development"]
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["JavaScript", "React", "Node.js", "MongoDB"]
          },
          "applicationDeadline": {
            "type": "string",
            "format": "date",
            "example": "2024-02-28"
          },
          "featured": {
            "type": "boolean",
            "default": false,
            "example": true
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "expired"],
            "default": "active",
            "example": "active"
          },
          "applicants": {
            "type": "integer",
            "example": 45
          },
          "views": {
            "type": "integer",
            "example": 320
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
      "Company": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tech Solutions Ltd"
          },
          "description": {
            "type": "string",
            "example": "Leading technology company in Rwanda"
          },
          "industry": {
            "type": "string",
            "example": "Information Technology"
          },
          "size": {
            "type": "string",
            "enum": ["startup", "small", "medium", "large", "enterprise"],
            "example": "medium"
          },
          "website": {
            "type": "string",
            "format": "uri",
            "example": "https://techsolutions.rw"
          },
          "logo": {
            "type": "string",
            "format": "uri",
            "example": "/uploads/companies/tech_solutions_logo.png"
          }
        }
      },
      "Employment": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["full-time", "part-time", "contract", "internship", "freelance"],
            "example": "full-time"
          },
          "category": {
            "type": "string",
            "enum": ["technology", "finance", "healthcare", "education", "engineering", "marketing", "sales", "other"],
            "example": "technology"
          },
          "level": {
            "type": "string",
            "enum": ["entry", "junior", "mid", "senior", "lead", "executive"],
            "example": "mid"
          },
          "remote": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "hybrid": {
            "type": "boolean",
            "default": false,
            "example": false
          }
        }
      },
      "JobLocation": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "KG 123 Street"
          },
          "city": {
            "type": "string",
            "example": "Kigali"
          },
          "province": {
            "type": "string",
            "example": "Kigali City"
          },
          "country": {
            "type": "string",
            "default": "Rwanda",
            "example": "Rwanda"
          }
        }
      },
      "Salary": {
        "type": "object",
        "properties": {
          "min": {
            "type": "number",
            "format": "float",
            "example": 800000
          },
          "max": {
            "type": "number",
            "format": "float",
            "example": 1200000
          },
          "currency": {
            "type": "string",
            "default": "RWF",
            "example": "RWF"
          },
          "period": {
            "type": "string",
            "enum": ["hour", "month", "year"],
            "default": "month",
            "example": "month"
          },
          "negotiable": {
            "type": "boolean",
            "default": false,
            "example": true
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "example": 1
          },
          "pages": {
            "type": "integer",
            "example": 5
          },
          "total": {
            "type": "integer",
            "example": 50
          },
          "limit": {
            "type": "integer",
            "example": 10
          }
        }
      }
    }
  },
  "paths": {
...
    "/universities": {
      "get": {
        "tags": ["Universities"],
        "summary": "Get all universities",
        "description": "Retrieve universities with filtering, search, and pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 12,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search query for university name or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "Filter by city",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "province",
            "in": "query",
            "description": "Filter by province",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "description": "Show only featured universities",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["name", "rating", "newest"],
              "default": "name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Universities retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "universities": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/University"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get all jobs",
        "description": "Retrieve jobs with filtering, search, and pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 12,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search query for job title or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by employment category",
            "schema": {
              "type": "string",
              "enum": ["technology", "finance", "healthcare", "education", "engineering", "marketing", "sales", "other"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by employment type",
            "schema": {
              "type": "string",
              "enum": ["full-time", "part-time", "contract", "internship", "freelance"]
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Filter by employment level",
            "schema": {
              "type": "string",
              "enum": ["entry", "junior", "mid", "senior", "lead", "executive"]
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "Filter by location",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "Filter by city",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "province",
            "in": "query",
            "description": "Filter by province",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "remote",
            "in": "query",
            "description": "Show only remote jobs",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "description": "Show only featured jobs",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["newest", "oldest", "salary", "title"],
              "default": "newest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/contact": {
      "post": {
        "tags": ["Contact"],
        "summary": "Submit contact form",
        "description": "Send a message through the contact form",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "message"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@example.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+250788123456"
                  },
                  "subject": {
                    "type": "string",
                    "example": "Question about universities"
                  },
                  "message": {
                    "type": "string",
                    "minLength": 10,
                    "example": "I am interested in learning more about universities in Rwanda..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Message sent successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get admin dashboard data",
        "description": "Retrieve statistics and data for admin dashboard",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statistics": {
                      "type": "object",
                      "properties": {
                        "totalUsers": {
                          "type": "integer"
                        },
                        "totalUniversities": {
                          "type": "integer"
                        },
                        "totalJobs": {
                          "type": "integer"
                        },
                        "totalApplications": {
                          "type": "integer"
                        }
                      }
                    },
                    "recentActivity": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}